# apis/swagger.yml
...
swagger: "2.0"
schemes:
  - "http"
  - "https"
produces:
  - "application/json"
  - "text/plain"
consumes:
  - "application/json"
  - "text/plain"
basePath: "/v1.24"
info:
  title: "Pouch Engine API"
  version: "1.24"
  description: |
    API is an HTTP API served by Pouch Engine.
produces:
        - "application/json"
paths:
...
  /containers/json:
    get:
      summary: "List containers"
      operationId: "ContainerList"
      produces: ["application/json"]
      responses:
        200:
          description: "Summary containers that matches the query"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Container"
        500:
          $ref: "#/responses/500ErrorResponse"
      parameters:
        - name: "all"
          in: "query"
          description: "Return all containers. By default, only running containers are shown"
          type: "boolean"
          default: false
...
  /containers/{id}:
    delete:
      summary: "Remove one container"
      operationId: "ContainerRemove"
      parameters:
        - $ref: "#/parameters/id"
        - name: "force"
          in: "query"
          description: "If the container is running, force query is used to kill it and remove it forcefully."
          type: "boolean"
      responses:
        204:
          description: "no error"
        404:
          $ref: "#/responses/404ErrorResponse"
        500:
          $ref: "#/responses/500ErrorResponse"
      tags: ["Container"]
...
definitions:
  Error:
    type: "object"
    properties:
      message:
        type: string
...
  HostConfig:
    description: "Container configuration that depends on the host we are running on"
    type: "object"
    allOf:
      - properties:
          # Applicable to all platforms
          Binds:
            type: "array"
            description: |
              A list of volume bindings for this container. Each volume binding is a string in one of these forms:

              - `host-src:container-dest` to bind-mount a host path into the container. Both `host-src`, and `container-dest` must be an _absolute_ path.
              - `host-src:container-dest:ro` to make the bind mount read-only inside the container. Both `host-src`, and `container-dest` must be an _absolute_ path.
              - `volume-name:container-dest` to bind-mount a volume managed by a volume driver into the container. `container-dest` must be an _absolute_ path.
              - `volume-name:container-dest:ro` to mount the volume read-only inside the container.  `container-dest` must be an _absolute_ path.
            items:
              type: "string"
          ContainerIDFile:
            type: "string"
            description: "Path to a file where the container ID is written"
          LogConfig:
            description: "The logging configuration for this container"
            type: "object"
            $ref: "#/definitions/LogConfig"
          RestartPolicy:
            type: "object"
            description: "Restart policy to be used to manage the container"
            $ref: "#/definitions/RestartPolicy"
          NetworkMode:
            type: "string"
            description: "Network mode to use for this container. Supported standard values are: `bridge`, `host`, `none`, and `container:<name|id>`. Any other value is taken as a custom network's name to which this container should connect to."
          PortBindings:
            type: "object"
            description: "A map of exposed container ports and the host port they should map to."
            $ref: "#/definitions/PortMap"
          AutoRemove:
            type: "boolean"
            description: "Automatically remove the container when the container's process exits. This has no effect if `RestartPolicy` is set."
          VolumeDriver:
            type: "string"
            description: "Driver that this container uses to mount volumes."
          VolumesFrom:
            type: "array"
            description: "A list of volumes to inherit from another container, specified in the form `<container name>[:<ro|rw>]`."
            items:
              type: "string"
          CapAdd:
            type: "array"
            description: "A list of kernel capabilities to add to the container."
            items:
              type: "string"
          CapDrop:
            type: "array"
            description: "A list of kernel capabilities to drop from the container."
            items:
              type: "string"
          Dns:
            type: "array"
            description: "A list of DNS servers for the container to use."
            items:
              type: "string"
          DnsOptions:
            type: "array"
            description: "A list of DNS options."
            items:
              type: "string"
          DnsSearch:
            type: "array"
            description: "A list of DNS search domains."
            items:
              type: "string"
          ExtraHosts:
            type: "array"
            description: |
              A list of hostnames/IP mappings to add to the container's `/etc/hosts` file. Specified in the form `["hostname:IP"]`.
            items:
              type: "string"
          GroupAdd:
            type: "array"
            description: "A list of additional groups that the container process will run as."
            items:
              type: "string"
          IpcMode:
            type: "string"
            description: |
                    IPC sharing mode for the container. Possible values are:
                    - `"none"`: own private IPC namespace, with /dev/shm not mounted
                    - `"private"`: own private IPC namespace
                    - `"shareable"`: own private IPC namespace, with a possibility to share it with other containers
                    - `"container:<name|id>"`: join another (shareable) container's IPC namespace
                    - `"host"`: use the host system's IPC namespace
                    If not specified, daemon default is used, which can either be `"private"`
                    or `"shareable"`, depending on daemon version and configuration.
          Cgroup:
            type: "string"
            description: "Cgroup to use for the container."
          Links:
            type: "array"
            description: "A list of links for the container in the form `container_name:alias`."
            items:
              type: "string"
          OomScoreAdj:
            type: "integer"
            description: |
                An integer value containing the score given to the container in order to tune OOM killer preferences.
                The range is in [-1000, 1000].
            type: "integer"
            format: "int"
            x-nullable: false
            minimum: -1000
            maximum: 1000
          PidMode:
            type: "string"
            description: |
              Set the PID (Process) Namespace mode for the container. It can be either:
              - `"container:<name|id>"`: joins another container's PID namespace
              - `"host"`: use the host's PID namespace inside the container
          Privileged:
            type: "boolean"
            description: "Gives the container full access to the host."
          PublishAllPorts:
            type: "boolean"
            description: "Allocates a random host port for all of a container's exposed ports."
          ReadonlyRootfs:
            type: "boolean"
            description: "Mount the container's root filesystem as read only."
          SecurityOpt:
            type: "array"
            description: "A list of string values to customize labels for MLS systems, such as SELinux."
            items:
              type: "string"
          StorageOpt:
            type: "object"
            description: |
              Storage driver options for this container, in the form `{"size": "120G"}`.
            additionalProperties:
              type: "string"
          Tmpfs:
            type: "object"
            description: |
              A map of container directories which should be replaced by tmpfs mounts, and their corresponding mount options. For example: `{ "/run": "rw,noexec,nosuid,size=65536k" }`.
            additionalProperties:
              type: "string"
          UTSMode:
            type: "string"
            description: "UTS namespace to use for the container."
          UsernsMode:
            type: "string"
            description: "Sets the usernamespace mode for the container when usernamespace remapping option is enabled."
          ShmSize:
            type: "integer"
            description: "Size of `/dev/shm` in bytes. If omitted, the system uses 64MB."
            minimum: 0
          Sysctls:
            type: "object"
            description: |
              A list of kernel parameters (sysctls) to set in the container. For example: `{"net.ipv4.ip_forward": "1"}`
            additionalProperties:
              type: "string"
          Runtime:
            type: "string"
            description: "Runtime to use with this container."
          # Applicable to Windows
          ConsoleSize:
            type: "array"
            description: "Initial console size, as an `[height, width]` array. (Windows only)"
            minItems: 2
            maxItems: 2
            items:
              type: "integer"
              minimum: 0
          Isolation:
            type: "string"
            description: "Isolation technology of the container. (Windows only)"
            enum:
              - "default"
              - "process"
              - "hyperv"
          EnableLxcfs:
            description: "Whether to enable lxcfs."
            type: "boolean"
            x-nullable: false
          Rich:
            type: "boolean"
            description: "Whether to start container in rich container mode. (default false)"
            x-nullable: false
          RichMode:
            type: "string"
            description: "Choose one rich container mode.(default dumb-init)"
            enum:
             - "dumb-init"
             - "sbin-init"
             - "systemd"
          InitScript:
            type: "string"
            description: "Initial script executed in container. The script will be executed before entrypoint or command"
      - $ref: "#/definitions/Resources"
...
  LogConfig:
    description: "The logging configuration for this container"
    type: "object"
    properties:
      Type:
        type: "string"
        x-go-name: "LogDriver"
        enum:
          - "json-file"
          - "syslog"
          - "journald"
          - "gelf"
          - "fluentd"
          - "awslogs"
          - "splunk"
          - "etwlogs"
          - "none"
      Config:
        type: "object"
        x-go-name: "LogOpts"
        additionalProperties:
          type: "string"
...
  Container:
    description: |
      an array of Container contains response of Engine API:
      GET "/containers/json"
    type: "object"
    properties:
      Id:
        description: "Container ID"
        type: "string"
      Names:
        type: "array"
        items:
          type: "string"
        example:
          - "container_1"
          - "container_2"
      Image:
        type: "string"
      ImageID:
        type: "string"
      Command:
        type: "string"
      Created:
        description: "Created time of container in daemon."
        type: "integer"
        format: "int64"
      SizeRw:
        type: "integer"
        format: "int64"
      SizeRootFs:
        type: "integer"
        format: "int64"
      Labels:
        type: "object"
        additionalProperties:
          type: "string"
      State:
        type: "string"
      Status:
        type: "string"
      HostConfig:
        description: |
          In Moby's API, HostConfig field in Container struct has following type 
          struct { NetworkMode string `json:",omitempty"` }
          In Pouch, we need to pick runtime field in HostConfig from daemon side to judge runtime type,
          So Pouch changes this type to be the complete HostConfig.
          Incompatibility exists, ATTENTION.
        $ref: "#/definitions/HostConfig"
        x-nullable: false
      Mounts:
        type: "array"
        description: "Set of mount point in a container."
        items:
          $ref: "#/definitions/MountPoint"
      NetworkSettings:
        type: "object"
        properties:
          Networks:
            additionalProperties:
              $ref: "#/definitions/EndpointSettings"
              x-nullable: true
...
  EndpointSettings:
    description: "Configuration for a network endpoint."
    type: "object"
    properties:
      # Configurations
      IPAMConfig:
        $ref: "#/definitions/EndpointIPAMConfig"
        x-nullable: true
      Links:
        type: "array"
        items:
          type: "string"
        example:
          - "container_1"
          - "container_2"
      Aliases:
        type: "array"
        items:
          type: "string"
        example:
          - "server_x"
          - "server_y"

      # Operational data
      NetworkID:
        description: |
          Unique ID of the network.
        type: "string"
        example: "08754567f1f40222263eab4102e1c733ae697e8e354aa9cd6e18d7402835292a"
      EndpointID:
        description: |
          Unique ID for the service endpoint in a Sandbox.
        type: "string"
        example: "b88f5b905aabf2893f3cbc4ee42d1ea7980bbc0a92e2c8922b1e1795298afb0b"
      Gateway:
        description: |
          Gateway address for this network.
        type: "string"
        example: "172.17.0.1"
      IPAddress:
        description: |
          IPv4 address.
        type: "string"
        example: "172.17.0.4"
      IPPrefixLen:
        description: |
          Mask length of the IPv4 address.
        type: "integer"
        example: 16
      IPv6Gateway:
        description: |
          IPv6 gateway address.
        type: "string"
        example: "2001:db8:2::100"
      GlobalIPv6Address:
        description: |
          Global IPv6 address.
        type: "string"
        example: "2001:db8::5689"
      GlobalIPv6PrefixLen:
        description: |
          Mask length of the global IPv6 address.
        type: "integer"
        format: "int64"
        example: 64
      MacAddress:
        description: |
          MAC address for the endpoint on this network.
        type: "string"
        example: "02:42:ac:11:00:04"
      DriverOpts:
        description: |
          DriverOpts is a mapping of driver options and values. These options
          are passed directly to the driver and are driver specific.
        type: "object"
        x-nullable: true
        additionalProperties:
          type: "string"
        example:
          com.example.some-label: "some-value"
          com.example.some-other-label: "some-other-value"
  EndpointIPAMConfig:
    description: "IPAM configurations for the endpoint"
    type: "object"
    properties:
      IPv4Address:
        description: "ipv4 address"
        type: "string"
      IPv6Address:
        description: "ipv6 address"
        type: "string"
      LinkLocalIPs:
        description: "link to the list of local ip"
        type: "array"
        x-nullable: false
        items:
          type: "string"
...
  MountPoint:
    type: "object"
    description: "A mount point inside a container"
    x-nullable: false
    properties:
      Type:
        type: "string"
      ID:
        type: "string"
      Name:
        type: "string"
      Source:
        type: "string"
      Destination:
        type: "string"
      Driver:
        type: "string"
      Mode:
        type: "string"
      RW:
        type: "boolean"
      CopyData:
        type: "boolean"
      Named:
        type: "boolean"
      Replace:
        type: "string"
      Propagation:
        type: "string"
...
  PortMap:
    description: |
      PortMap describes the mapping of container ports to host ports, using the
      container's port-number and protocol as key in the format `<port>/<protocol>`,
      for example, `80/udp`.

      If a container's port is mapped for both `tcp` and `udp`, two separate
      entries are added to the mapping table.
    type: "object"
    additionalProperties:
      type: "array"
      items:
        $ref: "#/definitions/PortBinding"
    example:
      "443/tcp":
        - HostIp: "127.0.0.1"
          HostPort: "4443"
      "80/tcp":
        - HostIp: "0.0.0.0"
          HostPort: "80"
        - HostIp: "0.0.0.0"
          HostPort: "8080"
      "80/udp":
        - HostIp: "0.0.0.0"
          HostPort: "80"
      "2377/tcp": null
  PortBinding:
    description: "PortBinding represents a binding between a host IP address and a host port"
    type: "object"
    x-nullable: true
    properties:
      HostIp:
        description: "Host IP address that the container's port is mapped to."
        type: "string"
        example: "127.0.0.1"
      HostPort:
        description: "Host port number that the container's port is mapped to."
        type: "string"
        example: "4443"
  RestartPolicy:
    description: "Define container's restart policy"
    type: "object"
    properties:
      Name:
        type: "string"
      MaximumRetryCount:
        type: "integer"
...
parameters:
  id:
    name: id
    in: path
    required: true
    description: ID or name of the container
    type: string
...
responses:
...
  404ErrorResponse:
    description: An unexpected 404 error occurred.
    schema:
      $ref: "#/definitions/Error"
  500ErrorResponse:
    description: An unexpected server error occurred.
    schema:
      $ref: "#/definitions/Error"